Move shared onboarding experiment types into protocol.
