+++ vscode/src/services/AuthProviderSimplified.ts
@@ -2,8 +2,9 @@ import * as vscode from 'vscode'
 
 import { DOTCOM_URL } from '@sourcegraph/cody-shared/src/sourcegraph-api/environments'
 
+import { AuthMethod } from '../chat/protocol'
+
 import { AuthProvider } from './AuthProvider'
-import { AuthMethod } from './OnboardingExperiment'
 
 // An auth provider for simplified onboarding. This is a sidecar to AuthProvider
 // so we can deprecate the experiment later. AuthProviderSimplified only works
