+++ vscode/src/services/OnboardingExperiment.ts
@@ -1,7 +1,1 @@
-export type AuthMethod = 'dotcom' | 'github' | 'gitlab' | 'google'
-
-export enum OnboardingExperimentArm {
-    Classic,
-    Simplified,
-    Default = Classic,
-}
+// TODO: Something here.
